pyproject.toml
[build-system]
requires      = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "freeq"
version = "0.0.14"
description = "Zero-setup queue with e2e encryption for free in 1 line of code"
readme = "README.md"
authors = [
    { name = "Viktor Hronec", email = "zamr666@gmail.com" },
]
license = { file = "LICENSE" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]
keywords = ["queue", "easy", "message queue", "zero-setup"]
dependencies = [
    "async-timeout==4.0.3",
    "blinker==1.6.2",
    "certifi==2023.7.22",
    "cffi==1.15.1",
    "charset-normalizer==3.2.0",
    "click==8.1.7",
    "cryptography==41.0.3",
    "flask==2.3.3",
    "hiredis==2.2.3",
    "idna==3.4",
    "itsdangerous==2.1.2",
    "jinja2==3.1.2",
    "markupsafe==2.1.3",
    "pycparser==2.21",
    "pycryptodome==3.18.0",
    "redis[hiredis]==5.0.0",
    "requests==2.31.0",
    "urllib3==2.0.4",
    "werkzeug==2.3.7",
    "zstandard==0.21.0"
]
requires-python = ">=3.7"

[project.urls]
"Homepage" = "https://github.com/hronecviktor/freeq"
"Bug Tracker" = "https://github.com/hronecviktor/freeq/issues"
